backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
pydantic-settings==2.7.1
python3-saml==1.16.0
PyJWT==2.10.1
aiohttp==3.11.11
websockets==14.1
python-multipart==0.0.20
